Refactor header's css using scss syntax
Scss문법에 맞춰 헤더 관련 파일을 리팩토링함
