--- src/component/common/menu/style/dropdownMenu.scss
@@ -1,53 +1,0 @@
-@import url(http://fonts.googleapis.com/earlyaccess/notosanskr.css);
-
-.react-Dropdown {
-  width: 100%;
-  height: 100%;
-  background-color: #2c3e50;
-  font-family: "Noto Sans KR", sans-serif;
-  font-weight: normal;
-  z-index: 9999;
-  position: relative;
-}
-
-.react-Dropdown:hover {
-  background-color: #212f3d;
-}
-
-.react-Dropdown .btn {
-  height: 100%;
-  margin: 0 8px;
-  display: flex;
-  align-items: center;
-  justify-content: center;
-  z-index: 9999;
-  color: white;
-}
-
-.react-Dropdown > .content {
-  display: none;
-  position: absolute;
-  width: inherit;
-  line-height: 60px;
-  box-shadow: 2px 3px 3px 0px #2c3e50;
-  z-index: 1;
-  z-index: 9999;
-}
-
-.react-Dropdown > .content > .react-MenuItem {
-  width: 100%;
-  height: 60px;
-  line-height: 22px;
-  word-break: break-word;
-  text-align: center;
-}
-
-.react-Dropdown:hover > .content {
-  display: flex;
-  flex-direction: column;
-}
-
-.react-Dropdown.rightdrop > .content {
-  top: 0px;
-  left: 100%;
-}
