--- src/component/common/menu/style/menu.scss
@@ -1,35 +1,0 @@
-@import url(http://fonts.googleapis.com/earlyaccess/notosanskr.css);
-
-.react-Menu {
-  display: flex;
-  font-family: "Noto Sans KR", sans-serif;
-  z-index: 9999;
-}
-
-.react-Menu a {
-  display: flex;
-  height: 100%;
-}
-
-.react-MenuItem {
-  flex: 0 0 auto;
-  width: fit-content;
-  height: 100%;
-  background-color: #2c3e50;
-  color: white;
-  font-size: 16px;
-  border: none;
-  word-break: break-all;
-  cursor: pointer;
-}
-
-.react-MenuItem:hover {
-  background-color: #212f3d;
-}
-
-.react-Menu .name {
-  width: 100%;
-  margin: 0 8px;
-  align-self: center;
-  color: white;
-}
