--- src/style/Header.scss
@@ -1,16 +1,0 @@
-header {
-  width: 100%;
-  min-width: 1300px;
-  line-height: 65px;
-  height: 65px;
-  background-color: #2c3e50;
-  box-shadow: 2px 4px 3px 0px #34495e;
-  position: sticky;
-  top: 0;
-  z-index: 1;
-}
-
-header nav {
-  width: 1300px;
-  margin: 0 auto;
-}
