+++ src/style/Header/headerStyle.scss
@@ -1,0 +1,17 @@
+@import url("http://fonts.googleapis.com/earlyaccess/notosanskr.css");
+
+$headerIndex: 1;
+$headerHeight: 65px;
+$headerFont: "Noto Sans KR", sans-serif;
+$headerFontSize: 16px;
+$headerColor: white;
+$headerBGColor: #2c3e50;
+$headerShadow: 2px 4px 3px 0px $headerBGColor;
+$headerMargin: 0 8px;
+
+$headerNavMinWidth: 1300px;
+
+$headerMenuHoverColor: #212f3d;
+$headerDropDownMenuHeight: 60px;
+
+$headerLogoHeight: 52px;
