--- src/style/Logo.scss
@@ -1,15 +1,0 @@
-.Logo {
-  a {
-    height: 52px;
-    padding: 0 8px;
-    font-size: 18px;
-    font-weight: 400;
-    color: white;
-    display: block;
-
-    img {
-      width: 100px;
-      vertical-align: middle;
-    }
-  }
-}
